Make test suites exported by renaming all tests/*_test.go to tests/*.go and moving Test*Suite functions out to tests/all_test.go
