+++ tests/admin.go
