+++ tests/continue_as_new.go
