+++ tests/functional_test_base_suite.go
@@ -25,10 +25,7 @@
 package tests
 
 import (
-	"testing"
-
 	"github.com/stretchr/testify/require"
-	"github.com/stretchr/testify/suite"
 	"go.uber.org/fx"
 
 	"go.temporal.io/server/common/primitives"
@@ -72,8 +69,4 @@ func (s *FunctionalTestBaseSuite) TestWithFxOptionsForService() {
 
 func (s *FunctionalTestBaseSuite) SetupTest() {
 	s.Assertions = require.New(s.T())
-}
-
-func TestFunctionalTestBaseSuite(t *testing.T) {
-	suite.Run(t, new(FunctionalTestBaseSuite))
 }
