+++ tests/namespace_interceptor.go
